+++ R/req-perform-stream.R
@@ -29,7 +29,7 @@
 #'   TRUE
 #' }
 #' resp <- request(example_url()) |>
-#'   req_url_path("/stream-bytes/100000") |>
+#'   req_url_relative("/stream-bytes/100000") |>
 #'   req_perform_stream(show_bytes, buffer_kb = 32)
 #' resp
 req_perform_stream <- function(req,