+++ tests/testthat/test-iterate.R
@@ -1,6 +1,6 @@
 test_that("can perform multiple requests", {
   req <- request(example_url()) %>%
-    req_url_path("/iris") %>%
+    req_url_relative("/iris") %>%
     req_url_query(limit = 5)
 
   resps <- req_perform_iterative(
@@ -15,7 +15,7 @@ test_that("can perform multiple requests", {
 
 test_that("can save results to disk", {
   req <- request(example_url()) %>%
-    req_url_path("/iris") %>%
+    req_url_relative("/iris") %>%
     req_url_query(limit = 5)
 
   dir <- withr::local_tempdir()
@@ -33,7 +33,7 @@ test_that("can save results to disk", {
 
 test_that("user temination still returns data", {
   req <- request(example_url()) %>%
-    req_url_path("/iris") %>%
+    req_url_relative("/iris") %>%
     req_url_query(limit = 5)
   next_req <- function(resp, req) interrupt()
 
@@ -46,7 +46,7 @@ test_that("user temination still returns data", {
 
 test_that("can retrieve all pages", {
   req <- request(example_url()) %>%
-    req_url_path("/iris") %>%
+    req_url_relative("/iris") %>%
     req_url_query(limit = 1)
 
   i <- 1