+++ tests/testthat/test-req-perform-stream.R
@@ -1,5 +1,5 @@
 test_that("req_stream() is deprecated", {
-  req <- request(example_url()) %>% req_url_path("/stream-bytes/100")
+  req <- request(example_url()) %>% req_url_relative("/stream-bytes/100")
   expect_snapshot(
     resp <- req_stream(req, identity, buffer_kb = 32)
   )