+++ tests/testthat/test-req-url.R
@@ -3,44 +3,6 @@ test_that("can override url", {
   expect_equal(req_url(req, "http://foo.com:10")$url, "http://foo.com:10")
 })
 
-test_that("automatically adds /", {
-  req1 <- request("http://example.com")
-  req2 <- request("http://example.com/")
-
-  expect_equal(req_url_path(req1, "/index.html")$url, "http://example.com/index.html")
-  expect_equal(req_url_path(req1, "index.html")$url, "http://example.com/index.html")
-  expect_equal(req_url_path(req2, "/index.html")$url, "http://example.com/index.html")
-  expect_equal(req_url_path(req2, "index.html")$url, "http://example.com/index.html")
-
-  expect_equal(req_url_path_append(req1, "index.html")$url, "http://example.com/index.html")
-  expect_equal(req_url_path_append(req1, "/index.html")$url, "http://example.com/index.html")
-  expect_equal(req_url_path_append(req2, "index.html")$url, "http://example.com/index.html")
-  expect_equal(req_url_path_append(req2, "/index.html")$url, "http://example.com/index.html")
-})
-
-test_that("can append multiple components", {
-  req <- request("http://example.com/x")
-  expect_equal(req_url_path(req, "a", "b")$url, "http://example.com/a/b")
-  expect_equal(req_url_path_append(req, "a", "b")$url, "http://example.com/x/a/b")
-})
-
-test_that("can handle empty path", {
-  req <- request("http://example.com/x")
-  expect_equal(req_url_path(req)$url, "http://example.com/")
-  expect_equal(req_url_path_append(req)$url, "http://example.com/x")
-  expect_equal(req_url_path(req, NULL)$url, "http://example.com/")
-  expect_equal(req_url_path_append(req, NULL)$url, "http://example.com/x")
-
-  expect_equal(req_url_path(req, "")$url, "http://example.com/")
-  expect_equal(req_url_path_append(req, "")$url, "http://example.com/x")
-})
-
-test_that("can handle path vector", {
-  req <- request("http://example.com/x")
-  expect_equal(req_url_path(req, c("a", "b"))$url, "http://example.com/a/b")
-  expect_equal(req_url_path_append(req, c("a", "b"))$url, "http://example.com/x/a/b")
-  expect_equal(req_url_path_append(req, c("a", "b"), NULL)$url, "http://example.com/x/a/b")
-})
 
 test_that("can set query params", {
   req <- request("http://example.com/")